Cleanup threading, enable optional use of pthread functions
Simplify the threading, queueing and locking logic.
Also add the possibility to use plain pthread functions instead
of C++11 std::thread. This is a workaround for a broken libstc++
on some platforms, leading to a 'pure virtual function called'
error. The workaround can be enabled by calling CMake with the
argument '-DUSE_PTHREAD=1'

Signed-off-by: warped-rudi <r.ihle@s-t.de>